feat(folder): filter out bound folders in `SelectFolder` component
There are two ways to tell user that "you have already bind this song to
these folders so you don't have to bind them again", one is checking it
before user "see" it, while the other way is to directly filter it out.

The reason that using the second way is the `Bind API` is implemented as
`incremental update`. Which means, you can't unbind a folder by
unchecking folders and press submit.The only way to remove binding is
pressing `delete` button in folder management page.