fix(database): OOM error is returned when reading beatoraja databasement file
Problem: One user reports that `Out of memory` error is returned at
`initialize` when trying to initialize `LampGhost`. This is because
`Sqlite` would try to initialize databasement file if the file is not
exist. If user provides a directory, obviously `Sqlite` can never open
it and never initialize the correct
It would keep trying until limit is reached.

This commit optimize two things:

- Force `LampGhost` connecting local databasement file with `READ_ONLY`
  flag.
- Pre-check before actually open the databasement file.