feat(rivalTag): add `symbol` feature
Show a symbol when importing a rival, which has locked version into
beatoraja. For example, if there was a rival named 'Catiz', and its
version has been locked to `sl4 First Clear`(which is a rival tag data
in lampghost). When importing into beatoraja, it would be cool if it's
rendered as `Catiz(sl4)`.

- Add a new field `symbol` for table `rival_tag`, which represents the
  'symbol' of the tag
- When generating rival data, if related rival tag data has symbol,
  applying it as part of the rival's name